requirements.txt
PySide6
folium
gpxpy
pandas
geopandas
Shapely
Pillow
exifread
mutagen
PyExifTool
matplotlib
pyproj
simplekml
